+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.2-dev
 
+- Fix: `BuildContext.ancestors` did not include `Component`s.
+
 ## 0.1.1
 
 - Fix: root component was not added to ancestor list.